+++ lib/clients/copyFactory/copyFactory_models.py
@@ -158,9 +158,12 @@ class CopyFactoryStrategySubscription(TypedDict):
     maxLeverage: Optional[float]
     """Optional setting indicating maximum leverage allowed when opening a new positions. Any trade which results in a
     higher leverage will be discarded."""
-    symbolMapping: List[CopyFactoryStrategySymbolMapping]
+    symbolMapping: Optional[List[CopyFactoryStrategySymbolMapping]]
     """Defines how symbol name should be changed when trading (e.g. when broker uses symbol names with unusual
     suffixes). By default this setting is disabled and the trades are copied using signal source symbol name."""
+    tradeSizeScalingMode: Optional[str]
+    """If set to balance, the trade size on strategy subscriber will be scaled according to balance to preserve risk.
+    If value is none, than trade size will be preserved irregardless of the subscriber balance. Default is balance."""
 
 
 class CopyFactoryAccountUpdate(TypedDict):
@@ -288,9 +291,12 @@ class CopyFactoryStrategyUpdate(TypedDict):
     timeSettings: Optional[CopyFactoryStrategyTimeSettings]
     """Settings to manage copying timeframe and position lifetime. Default is to copy position within 1 minute from
     being opened at source and let the position to live for up to 90 days."""
-    symbolMapping: List[CopyFactoryStrategySymbolMapping]
+    symbolMapping: Optional[List[CopyFactoryStrategySymbolMapping]]
     """Defines how symbol name should be changed when trading (e.g. when broker uses symbol names with unusual
     suffixes). By default this setting is disabled and the trades are copied using signal source symbol name."""
+    tradeSizeScalingMode: Optional[str]
+    """If set to balance, the trade size on strategy subscriber will be scaled according to balance to preserve risk.
+    If value is none, than trade size will be preserved irregardless of the subscriber balance. Default is balance."""
 
 
 class CopyFactorySubscriberOrProvider(TypedDict):
@@ -419,9 +425,12 @@ class CopyFactoryPortfolioMember(TypedDict):
     maxLeverage: Optional[float]
     """Optional max leverage risk restriction. All trades resulting in a leverage value higher than specified will be
     skipped."""
-    symbolMapping: List[CopyFactoryStrategySymbolMapping]
+    symbolMapping: Optional[List[CopyFactoryStrategySymbolMapping]]
     """Defines how symbol name should be changed when trading (e.g. when broker uses symbol names with unusual
     suffixes). By default this setting is disabled and the trades are copied using signal source symbol name."""
+    tradeSizeScalingMode: Optional[str]
+    """If set to balance, the trade size on strategy subscriber will be scaled according to balance to preserve risk.
+    If value is none, than trade size will be preserved irregardless of the subscriber balance. Default is balance."""
 
 
 class CopyFactoryPortfolioStrategyUpdate(TypedDict):
@@ -461,9 +470,12 @@ class CopyFactoryPortfolioStrategyUpdate(TypedDict):
     maxLeverage: Optional[float]
     """Optional max leverage risk restriction. All trades resulting in a leverage value higher than specified will be
     skipped."""
-    symbolMapping: List[CopyFactoryStrategySymbolMapping]
+    symbolMapping: Optional[List[CopyFactoryStrategySymbolMapping]]
     """Defines how symbol name should be changed when trading (e.g. when broker uses symbol names with unusual
     suffixes). By default this setting is disabled and the trades are copied using signal source symbol name."""
+    tradeSizeScalingMode: Optional[str]
+    """If set to balance, the trade size on strategy subscriber will be scaled according to balance to preserve risk.
+    If value is none, than trade size will be preserved irregardless of the subscriber balance. Default is balance."""
 
 
 class CopyFactoryPortfolioStrategy(CopyFactoryPortfolioStrategyUpdate):